+++ eoreader/__meta__.py
@@ -17,4 +17,4 @@
 **EOReader** library
 """
 
-__version__ = "0.22.3"
+__version__ = "0.22.4"