+++ CHANGES.md
@@ -3,7 +3,7 @@
 ## 0.22.4 (2025-05-05)
 
 - DOC: Fix jupyter cache for documentation notebook [#211](https://github.com/sertit/eoreader/issues/211)
-- FIX: Raise missing file not found exception
+- FIX: Raise missing file not found exception - by @emmanuel-ferdman
 
 ## 0.22.3 (2025-05-05)