+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 0.22.4 (2025-mm-dd)
+## 0.22.4 (2025-07-07)
 
 - FIX: Add workaround for `T` vs `T1+T2` abstraction [issue in Awesome Spectral Index](https://github.com/awesome-spectral-indices/awesome-spectral-indices/issues/74) [#214](https://github.com/sertit/eoreader/issues/214)
 - FIX: Raise missing file not found exception - by @emmanuel-ferdman